FIX: Handle custom copy mode mappings by letting them integrate with copycat on their own
The bind / unbind of copy mode mappings does not work if a custom global mapping for that key already exists. For me, this resulted in copycat mode never being quit (regardless of how copy mode was quit, even through the built-in Enter it showed this), and as a result n/N always jumped to a match instead of being inserted literally.
Let custom copy mode mappings integrate with tmux-copycat on their own; i.e. by explicitly invoking copycat_mode_quit.sh as part of their binding. Extend copycat_quit_copy_mode_keys() to skip those bindings that do this, by passing the target script into the function and filtering on that.