FIX: No abort when no matches
The comments in copycat_mode_start.sh mention that copycat_generate_results.sh will exit 0 if no results, but that is a separate script now (and isn't sourced), so it cannot abort the current shell's processing.
The script actually needs to indicate an error via exit status 1, and copycat_mode_start.sh's main() can then "|| return 0" on that.